+++ packages/framework/bundle/request-map/bundle.js
@@ -65,9 +65,6 @@ function() {
 			"src" : "../../../../sources/framework/request/common/remove-map-layer-request.js"
 		}, {
 			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/request/common/show-map-layer-info-request.js"
-		}, {
-			"type" : "text/javascript",
 			"src" : "../../../../sources/framework/request/common/ctrl-key-down-request.js"
 		}, {
 			"type" : "text/javascript",