+++ sources/framework/core/core.js
@@ -117,10 +117,6 @@
                 this._handleRemoveMapLayerRequest(request);
                 return true;
             },
-            'ShowMapLayerInfoRequest': function (request) {
-                this._handleShowMapLayerInfoRequest(request);
-                return true;
-            },
             'RearrangeSelectedMapLayerRequest': function (request) {
                 this._handleRearrangeSelectedMapLayerRequest(request);
                 return true;