+++ ReleaseNotes.md
@@ -5,7 +5,11 @@
 ### Oskari 2.0 preparations
 
 - Oskari.mapframework.sandbox.Sandbox has been renamed Oskari.Sandbox
+
+#### Request/Event changes
 - Moved MapMoveRequest and AfterMapMoveEvent file from under sources to mapmodule
+- ShowMapLayerInfoRequest moved from under sources to backendstatus as it is bundle specific request
+- AfterShowMapLayerInfoEvent removed as backendstatus was the only user and it can react to request without the event.
 
 #### Marker handling changes
 - AfterHideMapMarkerEvent was removed as it's no longer used and is misleading as it was used to notify markerlayer being hidden.